Feature/endpoints actualizados

+++ backend/app/routes/resenias.py
@@ -5,40 +5,67 @@ resenias = Blueprint("resenias", __name__)
 
 @resenias.route("/resenias", methods = ["GET"]) # cliente y admin
 def ver_resenias():
-    conn = get_connection() 
-    cursor = conn.cursor() 
-
-    cursor.execute("SELECT * FROM resenias")
-    resenias = cursor.fetchall()
-
-    cursor.close()
-    conn.close()
-
-    return jsonify(resenias), 200
+     try:
+        conn = get_connection() 
+        cursor = conn.cursor() 
+    
+        cursor.execute("SELECT * FROM resenias")
+        resenias = cursor.fetchall()
+        return jsonify(resenias), 200
+        
+    except Exception as e:
+        return jsonify({"error": f"Error al obtener resenias: {str(e)}"}, 500
+                       
+    finally:
+        cursor.close()
+        conn.close()
 
 @resenias.route("/resenias", methods=["POST"])  # cliente 
-def agregar_resenia():
-    conn = get_connection()
-    cursor = conn.cursor()
-    data = request.get_json() 
-
-    cursor.execute("""INSERT INTO resenias (descripcion) VALUES (%s)""", (data["descripcion"],))
-    conn.commit()
-
-    cursor.close()
-    conn.close()
-
-    return jsonify({"mensaje": "Reseña agregada correctamente"}), 201
+def agregar_resenia(): 
+     try:
+        conn = get_connection()
+        cursor = conn.cursor()
+        data = request.get_json() 
+         
+        if data is None or "descripcion" not in data:
+            return jsonify({"error": f"Falta el campo requerido: descripcion"}), 400
+            
+        cursor.execute("""INSERT INTO resenias (descripcion) VALUES (%s)""", (data["descripcion"],))
+        conn.commit()
+         
+        cursor.execute("SELECT * FROM resenias WHERE descripcion = %s", (data["descripcion"],) )
+        resenia_creada = cursor.fetchone()
+        return jsonify(resenia_creada), 201
+         
+    except Exception as e:
+        return jsonify({"error": f"Error al agregar resenia: {str(e)}"}, 500
+                       
+    finally:
+        cursor.close()
+        conn.close()
 
 @resenias.route("/resenias/<int:id_resenias>", methods = ["DELETE"]) # admin 
 def eliminar_resenias(id_resenias):
-    conn = get_connection() 
-    cursor = conn.cursor() 
+    try:
+        conn = get_connection() 
+        cursor = conn.cursor() 
+          
+        cursor.execute("SELECT * FROM resenias WHERE id_resenias = %s", (id_resenias,))
+        resenia = cursor.fetchone()
 
-    cursor.execute("DELETE FROM resenias where id_resenias=%s", (id_resenias,)) 
-    conn.commit()
+        if resenia is None:
+            return jsonify({"error": f"No existe una resenia con el id : {id_resenias}"}), 404
+    
+        cursor.execute("DELETE FROM resenias where id_resenias=%s", (id_resenias,)) 
+        conn.commit()
 
-    cursor.close()
-    conn.close()
+        return jsonify(resenia), 200
 
-    return jsonify({"mensaje": "Reseña eliminada correctamente"}), 200
+    except Exception as e:
+        return jsonify({"error": f"Error al eliminar resenia: {str(e)}"}, 500
+                       
+    finally:
+        cursor.close()
+        conn.close()
+    
+
